refactoring

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -6,7 +6,6 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateCategoryRequest;
 use App\Http\Requests\EditCategoryRequest;
 use App\Models\Category;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
@@ -25,17 +24,13 @@ class CategoryController extends Controller
 
     public function store(CreateCategoryRequest $request)
     {
+        $data = $request->validated();
         if ($request->has('image')) {
             $filename = time() . $request->file('image')->getClientOriginalName();
-            $request->file('image')->storeAs('categories', $filename, 'public');
+            $data['image'] = $request->file('image')->storeAs('categories', $filename, 'public');
         }
 
-        Category::create([
-            'code' => $request->code,
-            'name' => $request->name,
-            'description' => $request->description,
-            'image' => $filename ?? null,
-        ]);
+        Category::create($data);
 
         return redirect()->route('admin.categories.index')->with('message', 'Category has been created.');
     }
@@ -52,25 +47,25 @@ class CategoryController extends Controller
 
     public function update(EditCategoryRequest $request, Category $category)
     {
+        $data = $request->validated();
         if ($request->has('image')) {
             Storage::delete($category->image);
 
             $filename = time() . $request->file('image')->getClientOriginalName();
-            $path = $request->file('image')->storeAs('categories', $filename, 'public');
+            $data['image'] = $request->file('image')->storeAs('categories', $filename, 'public');
         }
 
-        $category->update([
-            'code' => $request->code,
-            'name' => $request->name,
-            'description' => $request->description,
-            'image' => $path ?? null,
-        ]);
+        $category->update($data);
 
         return redirect()->route('admin.categories.index')->with('message', 'Category has been updated.');
     }
 
     public function destroy(Category $category)
     {
+        if ($category->products()->count()) {
+            return redirect()->route('admin.categories.index')->with('message', 'Cannot delete, category has products.');
+        }
+
         $category->delete();
 
         return redirect()->route('admin.categories.index')->with('message', 'Category has been deleted.');
